Corregido error de Ecualizacion

+++ GUI_SI/src/imagen/ImageClass.java
@@ -395,7 +395,7 @@ public class ImageClass {
         int [] acHisto = this.getAcumulativeValues();
         
         for (int i=0;i<255;i++){
-            trans_table[i] = Math.max(0, Math.round((256/img_size)*acHisto[i])-1);
+            trans_table[i] = (int) Math.max(0, Math.round(((float)256/(float)img_size)*acHisto[i])-1);
             System.out.println(trans_table[i]);
         }
         return TRANSFORM(pixels,trans_table);
